refactor(codegen): miscellaneous cleanup, amend tests (#2519)
Simplify the generate_classes.py script. Remove the backup and exclude functionality — this can be done manually if needed in the future. Improve the console output by showing lists in columns.

Also, one of the old commits we tested had some funky variable definitions in gwf-evta.dfn, which I think we can safely assume are no longer part of the spec. This was breaking the codegen CI tests.