refactor(codegen): better workaround for legacy dfn format requirement (#2520)
Instead of re-reading DFNs from disk in a Jinja filter, attach the legacy DFN format to the context passed into the template for now. See #2385 (comment) for why we need to do such a thing in the first place. This step will let us remove DFNs from version control, coming shortly to follow.