fix(get-modflow): allow run_main bindir to be a path (#2560)
The run_main function downloads MOFLOW and related programs. A bindir argument has to be provided, of type str or Path according to the docstring. There was call to the .startswith method on line 383. However, a Path object does not have a startswith method. Fix run_main to accept bindir as either str or Path.