chore(dependencies): add upper bounds for numpy/pandas (#2662)
Add upper bounds to numpy and pandas in pyproject.toml, and remove --prerelease-allow from uv sync command in smoke tests. CI was failing because these two factors combined to allow prerelease versions of numpy and pandas, which broke some tests. The intent behind --allow-prerelease was to get a prerelease version of vtk on ARM macs, but seems a full version is now available on PyPI.

I think this might clear the way to switch CI to pixi if we want, as described in #2424